Improve 'gitbox' help information

+++ scripts/gitbox/lib/gitbox.setup.zsh
@@ -1,9 +1,89 @@
-function usage-setup() {
-    printf "'${YELLOW}gitbox setup${NORMAL}' provides several utilities to configure git\n"
-    printf "${YELLOW}%s${NORMAL}\n" "usage: gitbox setup [aliases | attributes | user | email | -h, --help]"
+source "${SANDBIN_HOME}/scripts/lib/usage.lib.zsh"
+source "${SANDBIN_HOME}/scripts/lib/colors.lib.zsh"
+source "${SANDBIN_HOME}/scripts/lib/git-functions.lib.zsh"
+
+function usage_setup() {
+    local mode=$(usage_mode $1)
+    local color=$(usage_description_color $mode $2)
+
+    $(usage_show_description $mode) && printf "${color}Provides several utilities to configure git${NORMAL}\n"
+    $(usage_show_usage $mode) && printf "usage: ${BOLD}gitbox setup${NORMAL} [aliases | attributes | user | email] [-h, --help]\n"
+
+    if usage_show_detailed $mode; then
+        printf "\n"
+        printf "    ${BOLD}aliases${NORMAL}         %s\n" "$(usage_setup_aliases description)"
+        printf "    ${BOLD}attributes${NORMAL}      %s\n" "$(usage_setup_gitattributes description)"
+        printf "    ${BOLD}user${NORMAL}            %s\n" "$(usage_setup_attribute user '' description)"
+        printf "    ${BOLD}email${NORMAL}           %s\n" "$(usage_setup_attribute email '' description)"
+        printf "\nOptions:\n"
+        printf "    ${BOLD}-h, --help${NORMAL}      Display this help\n"
+    fi
 }
 
-function gitbox-setup() {
+function usage_setup_aliases() {
+    local mode=$(usage_mode $1)
+    local color=$(usage_description_color $mode $2)
+
+    $(usage_show_description $mode) && printf "${color}Setup git aliases for a scope${NORMAL}\n"
+    $(usage_show_usage $mode) && printf "usage: ${BOLD}gitbox setup aliases${NORMAL} [--system | --global | --local] [-c, --clean-aliases] [<aliases file prefix>] [-h, --help]\n"
+
+    if usage_show_detailed $mode; then
+        printf "\n"
+        printf "    ${BOLD}<aliases file prefix>${NORMAL}       %s\n" "Alias definition file prefix (defaults to default)"
+        printf "\nOptions:\n"
+        printf "    ${BOLD}--system${NORMAL}                    %s\n" "Setup in system scope (all users)"
+        printf "    ${BOLD}--global${NORMAL}                    %s\n" "Setup in global scope (all user workspaces)"
+        printf "    ${BOLD}--local${NORMAL}                     %s\n" "Setup in local scope (only for current workspace)"
+        printf "    ${BOLD}-c, --clean-aliases${NORMAL}         %s\n" "First clean current scope aliases"
+        printf "    ${BOLD}-h, --help${NORMAL}                  %s\n" "Display this help"
+    fi
+}
+
+function usage_setup_gitattributes() {
+    local mode=$(usage_mode $1)
+    local color=$(usage_description_color $mode $2)
+
+    $(usage_show_description $mode) && printf "${color}Setup .gitattributes file in a git repository${NORMAL}\n"
+    $(usage_show_usage $mode) && printf "usage: ${BOLD}gitbox setup attributes${NORMAL}  [--to-dir <directory>] [-F | --force] [<attributes file prefix>] [-h | --help]\n"
+
+    if usage_show_detailed $mode; then
+        printf "\n"
+        printf "    ${BOLD}<attributes file prefix>${NORMAL}    %s\n" ".gitattributes difinition file prefix (defaults to default)"
+        printf "\nOptions:\n"
+        printf "    ${BOLD}--to-dir <directory>${NORMAL}        %s\n" "Target directory to setup .gitattributes"
+        printf "    ${BOLD}-F, --force${NORMAL}                 %s\n" "Replace .gitattributes file if already exists"
+        printf "    ${BOLD}-h, --help${NORMAL}                  %s\n" "Display this help"
+    fi
+}
+
+function usage_setup_attribute() {
+    local attribute="$1"
+    local attribute_name="$2"
+    local mode=$(usage_mode $3)
+    local color=$(usage_description_color $mode $4)
+
+    $(usage_show_description $mode) && printf "${color}Setup '%s' git attribute${NORMAL}\n" "$attribute"
+    $(usage_show_usage $mode) && printf "usage: ${BOLD}gitbox setup %s${NORMAL}  [--system | --global | --local] <%s> [-h | --help]\n" "$attribute_name" "$attribute_name"
+
+    if usage_show_detailed $mode; then
+        printf "\n"
+        printf "    ${BOLD}<%s>${NORMAL}              %s\n" "$attribute_name" "Value for $attribute attribute"
+        printf "\nOptions:\n"
+        printf "    ${BOLD}--system${NORMAL}            %s\n" "Setup in system scope (all users)"
+        printf "    ${BOLD}--global${NORMAL}            %s\n" "Setup in global scope (all user workspaces)"
+        printf "    ${BOLD}--local${NORMAL}             %s\n" "Setup in local scope (only for current workspace)"
+        printf "    ${BOLD}-h, --help${NORMAL}          %s\n" "Display this help"
+    fi
+}
+
+
+function gitbox_setup() {
+
+    if [ $# = 0 ]; then
+        printf "${RED}gitbox setup: Ouch! There is not enough parameters 'gitbox setup [subcommand]'${NORMAL}\n"
+        usage_setup
+        exit 1
+    fi
 
     while [[ $# > 0 ]];  do
         key="$1"
@@ -11,49 +91,40 @@ function gitbox-setup() {
         case $key in
             aliases)
                 shift
-                gitbox-setup-aliases "$@"
+                gitbox_setup_aliases "$@"
                 exit $?
             ;;
             attributes)
                 shift
-                gitbox-setup-attributes "$@"
+                gitbox_setup_gitattributes "$@"
                 exit $?
             ;;
             user)
                 shift
-                gitbox-setup-attribute "user.name" "$@"
+                gitbox_setup_attribute "user.name" "$key" "$@"
                 exit $?
             ;;
             email)
                 shift
-                gitbox-setup-attribute "user.email" "$@"
+                gitbox_setup_attribute "user.email" "$key" "$@"
                 exit $?
             ;;
             -h|--help)
-                usage-setup
+                usage_setup "help"
                 exit 0
             ;;
             *)
                 printf "${RED}%s${NORMAL}\n" "Ouch! Unknown option '$key'. Please try agan!"
-                usage-setup
+                usage_setup
                 exit 1
             ;;
         esac
 
         shift;
     done
-
-    printf "${RED}%s${NORMAL}\n" "Ouch! There is not enough parameters 'gitbox setup [subcommand]'"
-    usage-setup
-    exit 1
 }
 
-function gitbox-setup-aliases-usage() {
-    printf "%s\n" "usage: gitbox setup aliases <scope> [-c | --clean-aliases] [<aliases file prefix>] [-h | --help]"
-    printf "\t- ${BOLD}scope:${NORMAL} --system | --global | --local\n"
-}
-
-function gitbox-setup-aliases() {
+function gitbox_setup_aliases() {
 
     while [[ $# > 0 ]];  do
         key="$1"
@@ -66,7 +137,7 @@ function gitbox-setup-aliases() {
                 clean_aliases=$key
             ;;
             -h|--help)
-                gitbox-setup-aliases-usage
+                usage_setup_aliases "help"
                 return 0
             ;;
             *)
@@ -78,12 +149,14 @@ function gitbox-setup-aliases() {
     done
 
     if ! is_a_git_workspace; then
-        printf "${RED}Ouch! I need a git repository to work${NORMAL}\n"
+        printf "${RED}gitbox setup aliases: Ouch! I need a git repository to work${NORMAL}\n"
+        usage_setup_aliases
         return 1;
     fi
 
     if [ -z $scope ]; then
-        printf "${RED}Ouch! I need a scope (ie. --system, --global, --local) to setup aliases${NORMAL}\n"
+        printf "${RED}gitbox setup aliases: Ouch! I need a scope (ie. --system, --global, --local) to setup aliases${NORMAL}\n"
+        usage_setup_aliases
         return 1;
     fi
 
@@ -99,11 +172,7 @@ function gitbox-setup-aliases() {
     source "${SANDBIN_HOME}/dotfiles/gitaliases/${aliases_file_prefix}.gitaliases"
 }
 
-function gitbox-setup-attributes-usage() {
-    printf "%s\n" "usage: gitbox setup attributes [--to-dir <directory>] [-F | --force] [<attributes file prefix>] [-h | --help]"
-}
-
-function gitbox-setup-attributes() {
+function gitbox_setup_gitattributes() {
 
     while [[ $# > 0 ]];  do
         key="$1"
@@ -111,7 +180,7 @@ function gitbox-setup-attributes() {
         case $key in
             --to-dir)
                 if [ ! -d "$2" ]; then
-                    printf "${RED}Ouch! '%s' doesn't seem to be a valid directory.${NORMAL}\n" "$2"
+                    printf "${RED}gitbox setup attributes: Ouch! '%s' doesn't seem to be a valid directory.${NORMAL}\n" "$2"
                     exit 1
                 fi
                 directory="$2"
@@ -121,7 +190,7 @@ function gitbox-setup-attributes() {
                 force=true
             ;;
             -h|--help)
-                gitbox-setup-attributes-usage
+                usage_setup_gitattributes "help"
                 exit 0
             ;;
             *)
@@ -137,53 +206,48 @@ function gitbox-setup-attributes() {
     installation_path=$directory/.gitattributes
 
     if ! is_a_git_workspace $directory; then
-        printf "${RED}Ouch! I need a git repository to work${NORMAL}\n"
+        printf "${RED}gitbox setup attributes: Ouch! I need a git repository to work${NORMAL}\n"
         exit 1;
     fi
 
     if [ -z $attributes_file_prefix ]; then
         attributes_file_prefix="default"
     fi
+    template_path="$SANDBIN_HOME/dotfiles/gitattributes/$attributes_file_prefix.gitattributes"
+    if [ ! -f $template_path ]; then
+        printf "${RED}gitbox setup attributes: Ouch! I need a .gitattributes template and '%s' doesn't exists${NORMAL}\n" "$attributes_file_prefix.gitattributes"
+        exit 1
+    fi
 
     if [ ! -f "$installation_path" ] || [ $force ] ; then
 
         if [ -f "$installation_path" ]; then
-            printf "${YELLOW}Forcing to rewrite existing .gitattributes file ['%s'] (Saving a copy in %s)${NORMAL}\n" "$installation_path" "$installation_path.oldBackup"
+            printf "${YELLOW}Forcing to rewrite existing .gitattributes file ['%s'] (Saving a copy in %s)${NORMAL}\n" "${installation_path#./}" "${installation_path#./}.oldBackup"
             mv "$installation_path" "$installation_path.oldBackup"
         fi
 
-        printf "${GREEN}Creating .gitattributes file ['%s'] from %s.gitattributes template.${NORMAL}\n" "$installation_path" "$attributes_file_prefix"
-        cp "$SANDBIN_HOME/dotfiles/gitattributes/$attributes_file_prefix.gitattributes" "$installation_path"
+        printf "${GREEN}Creating .gitattributes file ['%s'] from %s.gitattributes template.${NORMAL}\n" "${installation_path#./}" "$attributes_file_prefix"
+        cp "$template_path" "$installation_path"
         return $?
     else
-        printf "The file ${RED}'%s'${NORMAL} already exists! (use -F,--force to rewrite)\n" "$installation_path"
+        printf "${RED}gitbox setup attributes: The file '%s' already exists! (use -F, --force to rewrite)${NORMAL}\n" "${installation_path#./}"
         return 1
     fi
 }
 
-function gitbox-setup-attribute-usage() {
-    local attribute="$1"
-    local scope="$2"
-
-    if [ -z $scope ]; then
-        printf "usage: gitbox setup %s <scope> <%s value>\n" "$attribute" "$attribute"
-        printf "\t- ${BOLD}scope:${NORMAL} --system | --global | --local\n"
-    else
-        printf "usage: gitbox setup %s %s <%s value>\n" "$attribute" "$scope" "$attribute"
-    fi
-}
-
-function gitbox-setup-attribute() {
+function gitbox_setup_attribute() {
 
     local attribute="$1"
+    local attribute_name="$2"
+
     shift;
-
+    shift;
     while [[ $# > 0 ]];  do
         key="$1"
 
         case $key in
             -h|--help)
-                gitbox-setup-attribute-usage $attribute $scope
+                usage_setup_attribute $attribute $attribute_name "help"
                 return 0;
             ;;
                 --system|--global|--local)
@@ -197,14 +261,14 @@ function gitbox-setup-attribute() {
     done
 
     if [ -z $value ] || [ -z $scope ]; then
-        printf "${RED}%s${NORMAL}\n" "Ouch! There is not enough parameters."
-        gitbox-setup-attribute-usage $attribute $scope
+        printf "${RED}gitbox setup $attribute_name: Ouch! There is not enough parameters.${NORMAL}\n"
+        usage_setup_attribute $attribute $attribute_name
         exit 1
     fi
 
     set_git_attribute "$attribute" "$scope" "$value"
     if (( $? )); then
-        printf "${RED}Ouch! '$BOLD%s$NORMAL$RED' has not been configured as $BOLD%s$NORMAL$RED in %s scope. It's '%s' a git repository?${NORMAL}\n" "$value" "$attribute" "$scope" "$(pwd)";
+        printf "${RED}gitbox setup $attribute_name: Ouch! '$BOLD%s$NORMAL$RED' has not been configured as $BOLD%s$NORMAL$RED in %s scope. It's '%s' a git repository?${NORMAL}\n" "$value" "$attribute" "$scope" "$(pwd)";
     else
         printf "${GREEN}Great! '$BOLD%s$NORMAL$GREEN' has been configured as $BOLD%s$NORMAL$GREEN in %s scope ${NORMAL}\n" "$(get_git_attribute $attribute $scope)" "$attribute" "$scope";
     fi
